// src/types/types.d.ts
interface PathwayProgress {
  pathwayId: ObjectId; // ID of the pathway this progress belongs to, NOT database ID of the progress itself
  finished: boolean;
  name: string;
  moduleProgress: ModuleProgress[];
}
interface ModuleProgress {
  moduleId: ObjectId;
  finished: boolean;
  name: string;
  contentProgress: ContentProgress[]; // Map between content ID and whether that content is finished
}
interface ContentProgress {
  finished: boolean;
  name: string;
  videoTime: number;
}
interface Pathway {
  _id?: ObjectId;
  name: string;
  modules: PathwayModule[];
  part: string;
  order: number;
  tags: string[];
}
interface Pathway_Db extends WithId<Document> {
  _id?: ObjectId;
  tags: string[];
  modules: ObjectId[]; // Module document IDs
  part: string;
  order: number;
  name: string;
}
interface PathwayModule {
  _id?: ObjectId;
  name: string;
  presetContent: PresetContent[];
  personalizedContent: PersonalizedContent[];
  tags: string[];
}
interface PathwayModule_Db extends WithId<Document> {
  _id?: ObjectId;
  name: string;
  presetContent: PresetContent[];
  tags: string[];
}
interface PresetContent {
  priority: number;
  name: string;
  type: string;
  url: string;
  content?: string;
}
interface PersonalizedContent extends WithId<Document> {
  _id?: ObjectId;
  moduleId: ObjectId;
  priority: number;
  tags: string[];
  name: string;
  type: string;
  url: string;
  content?: string;
}